Sonoff: rename device files
