+++ server/services/sonoff/models/light_arilux_lc01.js
