+++ server/services/sonoff/models/switch_sonoff_4ch.js
