+++ server/services/sonoff/models/switch_sonoff_basic.js
