+++ server/services/sonoff/models/switch_sonoff_dual.js
