+++ server/services/sonoff/models/switch_sonoff_dual_r2.js
