+++ server/test/services/sonoff/models/light_arilux_lc01.test.js
