+++ server/test/services/sonoff/models/switch_sonoff_4ch.test.js
