+++ server/test/services/sonoff/models/switch_sonoff_basic.test.js
