+++ server/test/services/sonoff/models/switch_sonoff_dual.test.js
