+++ server/test/services/sonoff/models/switch_sonoff_dual_r2.test.js
